AI-Provider/backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
pyodbc>=5.0.1
pymssql>=2.2.11

# LLM and AI
langchain>=0.1.4
langchain-community>=0.0.16
langchain-core>=0.1.16
ollama>=0.4.0

# Vector Store for RAG
chromadb>=1.0.0
sentence-transformers>=2.3.1
faiss-cpu>=1.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Logging and Monitoring
loguru>=0.7.2

# Data Processing
pandas>=2.1.4

# Ukrainian Language Support
langdetect>=1.0.9
tqdm>=4.66.1
rich>=13.7.0
